refactor: renamed styles.js to baseStyles.js
